gsm.qc: Good Statistical Monitoring Testing Suite

The Good Statistical Monitoring or 'gsm' R packages provides a framework for statistical data monitoring. This package provides all a testing suite for all the packages involved.

See also

Author

Maintainer: Jeremy Wildfire jwildfire@gmail.com

Authors:

Zelos Zhu zelos.zhu@atorusresearch.com

Other contributors:

Gilead Sciences [copyright holder]